# perf: improve graph build performance