--- a/tensorboard/plugins/graph/tf_graph/tf-graph-scene.html.ts
+++ b/tensorboard/plugins/graph/tf_graph/tf-graph-scene.html.ts
@@ -273,6 +273,11 @@ export const template = html`
 
     /* --- Node label --- */
 
+    #root,
+    .node {
+      will-change: transform;
+    }
+
     .node > text.nodelabel {
       cursor: pointer;
       fill: #444;