--- a/tensorboard/plugins/graph/tf_graph/tf-graph-scene.html.ts
+++ b/tensorboard/plugins/graph/tf_graph/tf-graph-scene.html.ts
@@ -14,6 +14,7 @@ limitations under the License.
 ==============================================================================*/
 import {html} from '@polymer/polymer';
 
+// Please keep node font-size/classnames in sync with tf-graph-common/common.ts
 export const template = html`
   <style>
     :host {