Background: since SVG text does not support CSS's
text-overflow:ellipsis, we manually measure the text dimension using
`SVGTextElement.prototype.getComputedTextLength`.

Problem: `getComputedTextLength` causes relayout/restyle which ends up
spending nearly 10ms per operation which is prohibited expensive. This
process is repeated with O(n) and causes larger graph rendering to take
~160s in a text set.

Fix: instead of using SVG's getComputedTextLength, we now use canvas and
its measureText. Do note that we need to know the fontSize and, to
prevent invoking `computedStyle` (which will invoke restyle), we
estimate the fontSize which may cause less than ideal layout.

After the change, we no longer see restyle or relayout in performance
profiling and saw the last render step ("build scene") reduced from 160s
to 700ms (tested with graph.pbtxt that is ~250 MiB).

Do note that we still spend considerable amount of time (80s) on "normalizing
names " which we intend to fix later.

Before:
![image](https://user-images.githubusercontent.com/2547313/102577175-59cb9980-40ac-11eb-92a0-53bdc0de3bc6.png)

After:
![image](https://user-images.githubusercontent.com/2547313/102581972-7bca1980-40b6-11eb-8c89-bb5d14ac8339.png)